requirements.txt
alembic==1.13.1
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.2
pandas==2.3.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
six==1.17.0
SQLAlchemy==2.0.43
types-requests==2.32.4.20250809
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
